% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_adol_chu9d}
\alias{add_adol_chu9d}
\title{Add adolescent CHU-9D}
\usage{
add_adol_chu9d(
  ds_tb,
  adol_chu9d_scrg_ds = make_adol_chu9d_scrg_ds(),
  cleanse_1L_chr = c("all", "none", "dims", "items", "itemdims"),
  dim_pfx_1L_chr = "chu9d_dim",
  item_pfx_1L_chr = "chu9d_item",
  scale_by_1L_dbl = 1.1059,
  unweighted_1L_chr = "chu9d_cml",
  weighted_1L_chr = "chu9d_utl"
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{adol_chu9d_scrg_ds}{PARAM_DESCRIPTION, Default: make_adol_chu9d_scrg_ds()}

\item{cleanse_1L_chr}{Cleanse (a character vector of length one), Default: c("all", "none", "dims", "items", "itemdims")}

\item{dim_pfx_1L_chr}{Dimension prefix (a character vector of length one), Default: 'chu9d_dim'}

\item{item_pfx_1L_chr}{Item prefix (a character vector of length one), Default: 'chu9d_item'}

\item{scale_by_1L_dbl}{Scale by (a double vector of length one), Default: 1.1059}

\item{unweighted_1L_chr}{Unweighted (a character vector of length one), Default: 'chu9d_cml'}

\item{weighted_1L_chr}{Weighted (a character vector of length one), Default: 'chu9d_utl'}
}
\value{
Dataset (a tibble)
}
\description{
add_adol_chu9d() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add adolescent chu-9d. The function returns Dataset (a tibble).
}
\keyword{internal}
