% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_aqol4d_scores}
\alias{add_aqol4d_scores}
\title{Add AQoL-4D scores}
\usage{
add_aqol4d_scores(
  object_xx,
  country_1L_chr = NA_character_,
  itm_labels_chr = character(0),
  itm_prefix_1L_chr = "aqol4d_q",
  keep_all_1L_lgl = F,
  scrg_dss_ls = NULL,
  tot_unwtd_var_nm_1L_chr = "aqol4d_unwtd_dbl",
  utl_var_nm_1L_chr = "aqol4d_utl_dbl"
)
}
\arguments{
\item{object_xx}{Object (an output object of multiple potential types)}

\item{country_1L_chr}{Country (a character vector of length one), Default: 'NA'}

\item{itm_labels_chr}{Item labels (a character vector), Default: character(0)}

\item{itm_prefix_1L_chr}{Item prefix (a character vector of length one), Default: 'aqol4d_q'}

\item{keep_all_1L_lgl}{Keep all (a logical vector of length one), Default: F}

\item{scrg_dss_ls}{Scoring datasets (a list), Default: NULL}

\item{tot_unwtd_var_nm_1L_chr}{Total unweighted variable name (a character vector of length one), Default: 'aqol4d_unwtd_dbl'}

\item{utl_var_nm_1L_chr}{Utility variable name (a character vector of length one), Default: 'aqol4d_utl_dbl'}
}
\value{
Object (an output object of multiple potential types)
}
\description{
add_aqol4d_scores() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add aqol-4d scores. The function returns Object (an output object of multiple potential types).
}
\keyword{internal}
