% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_item_totals}
\alias{add_item_totals}
\title{Add item totals}
\usage{
add_item_totals(
  ds_tb,
  domains_ls = NULL,
  domains_prefix_1L_chr = character(0),
  domain_tfmn_fn = identity,
  items_prefix_1L_chr,
  total_var_nm_1L_chr,
  total_tfmn_fn = identity,
  type_fn = as.integer
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{domains_ls}{Domains (a list), Default: NULL}

\item{domains_prefix_1L_chr}{Domains prefix (a character vector of length one), Default: character(0)}

\item{domain_tfmn_fn}{Domain transformation (a function), Default: identity}

\item{items_prefix_1L_chr}{Items prefix (a character vector of length one)}

\item{total_var_nm_1L_chr}{Total variable name (a character vector of length one)}

\item{total_tfmn_fn}{Total transformation (a function), Default: identity}

\item{type_fn}{Type (a function), Default: as.integer}
}
\value{
Dataset (a tibble)
}
\description{
add_item_totals() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add item totals. The function returns Dataset (a tibble).
}
\keyword{internal}
