% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_paid_totals}
\alias{add_paid_totals}
\title{Add paid totals}
\usage{
add_paid_totals(
  ds_tb,
  ctg_var_nm_1L_chr = "PAID_burnout_risk_lgl",
  dict_ctg_1L_chr = "PAID",
  dictionary_r3 = ready4use::ready4use_dictionary(),
  items_prefix_1L_chr = "paid_",
  total_var_nm_1L_chr = "PAID_total_dbl",
  what_1L_chr = "ds",
  ...
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{ctg_var_nm_1L_chr}{Category variable name (a character vector of length one), Default: 'PAID_burnout_risk_lgl'}

\item{dict_ctg_1L_chr}{Dictionary category (a character vector of length one), Default: 'PAID'}

\item{dictionary_r3}{Dictionary (a ready4 submodule), Default: ready4use::ready4use_dictionary()}

\item{items_prefix_1L_chr}{Items prefix (a character vector of length one), Default: 'paid_'}

\item{total_var_nm_1L_chr}{Total variable name (a character vector of length one), Default: 'PAID_total_dbl'}

\item{what_1L_chr}{What (a character vector of length one), Default: 'ds'}

\item{...}{Additional arguments}
}
\value{
Object (an output object of multiple potential types)
}
\description{
add_paid_totals() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add paid totals. The function returns Object (an output object of multiple potential types).
}
\keyword{internal}
