% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_scores}
\alias{add_scores}
\title{Add scores}
\usage{
add_scores(
  X_YouthvarsProfile,
  scoring_tb,
  label_ds_1L_lgl = T,
  what_1L_chr = c("merged", "list")
)
}
\arguments{
\item{X_YouthvarsProfile}{PARAM_DESCRIPTION}

\item{scoring_tb}{Scoring (a tibble)}

\item{label_ds_1L_lgl}{Label dataset (a logical vector of length one), Default: T}

\item{what_1L_chr}{What (a character vector of length one), Default: c("merged", "list")}
}
\value{
Object (an output object of multiple potential types)
}
\description{
add_scores() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add scores. The function returns Object (an output object of multiple potential types).
}
\keyword{internal}
