% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_aqol4d_dim_disv}
\alias{calculate_aqol4d_dim_disv}
\title{Calculate AQoL-4D dimension disvalue}
\usage{
calculate_aqol4d_dim_disv(domain_1L_chr, ds_tb, scrg_dss_ls = NULL)
}
\arguments{
\item{domain_1L_chr}{Domain (a character vector of length one)}

\item{ds_tb}{Dataset (a tibble)}

\item{scrg_dss_ls}{Scoring datasets (a list), Default: NULL}
}
\value{
an AQoL-4D dimension disvalue (a double vector)
}
\description{
calculate_aqol4d_dim_disv() is a Calculate function that performs a numeric calculation. Specifically, this function implements an algorithm to calculate aqol-4d dimension disvalue. The function returns an AQoL-4D dimension disvalue (a double vector).
}
\keyword{internal}
