% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_depict.R
\name{depict-ScorzProfile}
\alias{depict-ScorzProfile}
\alias{depict,ScorzProfile-method}
\title{Depict features of an instance of a class by generating a plot}
\usage{
\S4method{depict}{ScorzProfile}(
  x,
  heights_int = NA_integer_,
  plot_rows_cols_pair_int = NA_integer_,
  tfmn_1L_chr = "title",
  type_1L_chr = "item_by_time",
  y_label_1L_chr = "",
  var_idcs_int = NA_integer_,
  ...
)
}
\arguments{
\item{x}{An object of class ScorzProfile}

\item{heights_int}{Heights (an integer vector), Default: NA}

\item{plot_rows_cols_pair_int}{Plot rows columns pair (an integer vector), Default: NA}

\item{tfmn_1L_chr}{Transformation (a character vector of length one), Default: 'title'}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'item_by_time'}

\item{y_label_1L_chr}{Y label (a character vector of length one), Default: ''}

\item{var_idcs_int}{Variable idcs (an integer vector), Default: NA}

\item{...}{Additional arguments}
}
\value{
Plot (an output object of multiple potential types)
}
\description{
depict method applied to ScorzProfile
}
