% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_adol_chu9d_scrg_ds}
\alias{make_adol_chu9d_scrg_ds}
\title{Make adolescent CHU-9D scoring dataset}
\usage{
make_adol_chu9d_scrg_ds()
}
\value{
Adolescent CHU-9D scoring (a dataset)
}
\description{
make_adol_chu9d_scrg_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make adolescent chu-9d scoring dataset. The function is called for its side effects and does not return a value.
}
\keyword{internal}
