% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_aqol4d_dict}
\alias{make_aqol4d_dict}
\title{Make AQoL-4D dictionary}
\usage{
make_aqol4d_dict(
  prefix_1L_chr = "aqol4d_q",
  scrg_dss_ls = NULL,
  tot_unwtd_var_nm_1L_chr = "aqol4d_unwtd_dbl",
  utl_var_nm_1L_chr = "aqol4d_utl_dbl"
)
}
\arguments{
\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'aqol4d_q'}

\item{scrg_dss_ls}{Scoring datasets (a list), Default: NULL}

\item{tot_unwtd_var_nm_1L_chr}{Total unweighted variable name (a character vector of length one), Default: 'aqol4d_unwtd_dbl'}

\item{utl_var_nm_1L_chr}{Utility variable name (a character vector of length one), Default: 'aqol4d_utl_dbl'}
}
\value{
an AQoL-4D (ready4 s3 class defining a data dictionary tibble.)
}
\description{
make_aqol4d_dict() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make aqol-4d dictionary. The function is called for its side effects and does not return a value.
}
\keyword{internal}
