% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_aqol4d_domain_qs_lup}
\alias{make_aqol4d_domain_qs_lup}
\title{Make AQoL-4D domain questions lookup table}
\usage{
make_aqol4d_domain_qs_lup()
}
\value{
an AQoL-4D domain questions (a lookup table)
}
\description{
make_aqol4d_domain_qs_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make aqol-4d domain questions lookup table. The function returns an AQoL-4D domain questions (a lookup table).
}
\keyword{internal}
