% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_aqol4d_domains_ls}
\alias{make_aqol4d_domains_ls}
\title{Make AQoL-4D domains list}
\usage{
make_aqol4d_domains_ls(domain_qs_lup = NULL)
}
\arguments{
\item{domain_qs_lup}{Domain questions (a lookup table), Default: NULL}
}
\value{
Domains (a list)
}
\description{
make_aqol4d_domains_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make aqol-4d domains list. The function returns Domains (a list).
}
\keyword{internal}
