% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_aqol4d_item_disvalue_lup}
\alias{make_aqol4d_item_disvalue_lup}
\title{Make AQoL-4D item disvalue lookup table}
\usage{
make_aqol4d_item_disvalue_lup()
}
\value{
Item disvalue (a lookup table)
}
\description{
make_aqol4d_item_disvalue_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make aqol-4d item disvalue lookup table. The function returns Item disvalue (a lookup table).
}
\keyword{internal}
