% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_aqol4d_params_lup}
\alias{make_aqol4d_params_lup}
\title{Make AQoL-4D parameters lookup table}
\usage{
make_aqol4d_params_lup(domain_qs_lup = NULL)
}
\arguments{
\item{domain_qs_lup}{Domain questions (a lookup table), Default: NULL}
}
\value{
an AQoL-4D parameters (a lookup table)
}
\description{
make_aqol4d_params_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make aqol-4d parameters lookup table. The function returns an AQoL-4D parameters (a lookup table).
}
\keyword{internal}
