% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_aqol4d_scrg_dss_ls}
\alias{make_aqol4d_scrg_dss_ls}
\title{Make AQoL-4D scoring datasets list}
\usage{
make_aqol4d_scrg_dss_ls()
}
\value{
an AQoL-4D scoring datasets (a list)
}
\description{
make_aqol4d_scrg_dss_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make aqol-4d scoring datasets list. The function returns an AQoL-4D scoring datasets (a list).
}
\keyword{internal}
