% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_aqol6d_item_nms}
\alias{make_aqol6d_item_nms}
\title{Make Assessment of Quality of Life Six Dimension item names}
\usage{
make_aqol6d_item_nms()
}
\value{
Assessment of Quality of Life Six Dimension item names (a character vector)
}
\description{
make_aqol6d_item_nms() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make assessment of quality of life six dimension item names. The function returns Assessment of Quality of Life Six Dimension item names (a character vector).
}
\keyword{internal}
