% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_aqol6d_scrg_dict}
\alias{make_aqol6d_scrg_dict}
\title{Make Assessment of Quality of Life Six Dimension scoring dictionary}
\usage{
make_aqol6d_scrg_dict(
  prefix_1L_chr = "aqol6d_q",
  scrg_dss_ls = NULL,
  tot_unwtd_var_nm_1L_chr = "aqol6d_total_c",
  utl_var_nm_1L_chr = "aqol6d_total_w"
)
}
\arguments{
\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'aqol6d_q'}

\item{scrg_dss_ls}{Scoring datasets (a list), Default: NULL}

\item{tot_unwtd_var_nm_1L_chr}{Total unweighted variable name (a character vector of length one), Default: 'aqol6d_total_c'}

\item{utl_var_nm_1L_chr}{Utility variable name (a character vector of length one), Default: 'aqol6d_total_w'}
}
\value{
Assessment of Quality of Life Six Dimension (ready4 s3 class defining a data dictionary tibble.)
}
\description{
make_aqol6d_scrg_dict() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make assessment of quality of life six dimension scoring dictionary. The function is called for its side effects and does not return a value.
}
\keyword{internal}
