% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_aqol6d_scrg_dss}
\alias{make_aqol6d_scrg_dss}
\title{Make Assessment of Quality of Life Six Dimension scoring datasets}
\usage{
make_aqol6d_scrg_dss()
}
\value{
Assessment of Quality of Life Six Dimension scoring datasets (a list)
}
\description{
make_aqol6d_scrg_dss() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make assessment of quality of life six dimension scoring datasets. The function returns Assessment of Quality of Life Six Dimension scoring datasets (a list).
}
\keyword{internal}
