% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_eq5d_dict}
\alias{make_eq5d_dict}
\title{Make EQ5D dictionary}
\usage{
make_eq5d_dict()
}
\value{
EQ5D dictionary (a ready4 S3)
}
\description{
make_eq5d_dict() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make eq5d dictionary. The function returns EQ5D dictionary (a ready4 S3).
}
\keyword{internal}
