% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_filtered_profile}
\alias{make_filtered_profile}
\title{Make filtered profile}
\usage{
make_filtered_profile(
  X_ScorzProfile = ScorzProfile(),
  nbr_of_digits_1L_int = 2L,
  profile_chr = character(0),
  timepoint_1L_int = 1L
)
}
\arguments{
\item{X_ScorzProfile}{PARAM_DESCRIPTION, Default: ScorzProfile()}

\item{nbr_of_digits_1L_int}{Number of digits (an integer vector of length one), Default: 2}

\item{profile_chr}{Profile (a character vector), Default: character(0)}

\item{timepoint_1L_int}{Timepoint (an integer vector of length one), Default: 1}
}
\value{
X (A dataset and its associated dictionary, descriptive statistics and metadata.)
}
\description{
make_filtered_profile() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make filtered profile. The function is called for its side effects and does not return a value.
}
\keyword{internal}
