% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_manufacture.R
\name{manufacture-ScorzAqol6}
\alias{manufacture-ScorzAqol6}
\alias{manufacture,ScorzAqol6-method}
\alias{manufacture-ScorzEuroQol5}
\alias{manufacture,ScorzEuroQol5-method}
\alias{manufacture-ScorzProfile}
\alias{manufacture,ScorzProfile-method}
\title{Manufacture a new object}
\usage{
\S4method{manufacture}{ScorzAqol6}(x, what_1L_chr = "domains", ...)

\S4method{manufacture}{ScorzEuroQol5}(x, what_1L_chr = "domains", ...)

\S4method{manufacture}{ScorzProfile}(
  x,
  custom_args_ls = NULL,
  custom_fn = NULL,
  domain_pfxs_1L_chr = character(0),
  what_1L_chr = "domains",
  ...
)
}
\arguments{
\item{x}{An object of class ScorzProfile}

\item{what_1L_chr}{What (a character vector of length one), Default: 'domains'}

\item{...}{Additional arguments}

\item{custom_args_ls}{Custom arguments (a list), Default: NULL}

\item{custom_fn}{Custom (a function), Default: NULL}

\item{domain_pfxs_1L_chr}{Domain prefixes (a character vector of length one), Default: character(0)}
}
\value{
Object (an output object of multiple potential types)

Object (an output object of multiple potential types)

Object (an output object of multiple potential types)
}
\description{
manufacture method applied to ScorzAqol6

manufacture method applied to ScorzEuroQol5

manufacture method applied to ScorzProfile
}
