% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_exhibit.R
\name{exhibit-ScorzProfile}
\alias{exhibit-ScorzProfile}
\alias{exhibit,ScorzProfile-method}
\title{Exhibit (print to console) features of a ready4 framework module (or sub-module)}
\usage{
\S4method{exhibit}{ScorzProfile}(x, type_1L_chr = "ds", ...)
}
\arguments{
\item{x}{An object of class ScorzProfile}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'ds'}

\item{...}{Additional arguments}
}
\description{
exhibit method applied to ScorzProfile
}
