% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_aqol6d_scrg_dict}
\alias{get_aqol6d_scrg_dict}
\title{Get Assessment of Quality of Life Six Dimension scoring dictionary}
\usage{
get_aqol6d_scrg_dict()
}
\description{
get_aqol6d_scrg_dict() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get assessment of quality of life six dimension scoring dictionary. The function is called for its side effects and does not return a value.
}
\keyword{internal}
