% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_scorz.R, R/scorz-package.R
\docType{package}
\name{scorz-package}
\alias{scorz}
\alias{scorz-package}
\title{scorz: Score Multi-Attribute Utility Instruments}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Modules from the ready4 youth mental health economic model (https://www.ready4-dev.com/docs/model/) that can be used to summarise multiple questionnaire responses in a single index measure. This package uses publicly available scoring algorithms to calculate health utility from multi-attribute utility instruments and has been developed with the ready4 framework (https://www.ready4-dev.com/). This development version of the scorz package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors (matthew.hamilton1@monash.edu).

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Modules from the ready4 youth mental health economic model (https://www.ready4-dev.com/docs/model/) that can be used to summarise multiple questionnaire responses in a single index measure. This package uses publicly available scoring algorithms to calculate health utility from multi-attribute utility instruments and has been developed with the ready4 framework (https://www.ready4-dev.com/). This development version of the scorz package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors (matthew.hamilton1@monash.edu).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/scorz/}
  \item \url{https://github.com/ready4-dev/scorz}
  \item \url{https://www.ready4-dev.com/}
}


Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/scorz/}
  \item \url{https://github.com/ready4-dev/scorz}
  \item \url{https://www.ready4-dev.com/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton1@monash.edu} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Authors:
\itemize{
  \item Caroline Gao \email{caroline.gao@orygen.org.au} (\href{https://orcid.org/0000-0002-0987-2759}{ORCID})
}

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item Headspace [funder]
  \item National Health and Medical Research Council [funder]
}

}
\keyword{internal}
